CROSS VALIDATION IS STILL NOT REALLY WORKING 
Need to distribute the results so it makes sens

+++ include/Validation/Validation.hpp
@@ -16,14 +16,14 @@ Parameters : Prediction Model, DataSet
 Inputs     : BaggingModel, const DataSet
 Outputs    : double
 */
-std::tuple<double, double, double> compute_accuracy(BaggingModel &model,
+std::tuple<float, float, float> compute_accuracy(BaggingModel &model,
                                             const DataSet &data) {                                              
   std::vector<float> exact = data.get_Labels(); 
   std::vector<float> prediction = model.predict(data);
 
-  double error = MAE::apply(exact, prediction);
-  double percentage_Error = MAPE::apply(exact, prediction);
-  double standard_deviation = std::sqrt(RIV::apply(exact, prediction));
+  float error = MAE::apply(exact, prediction);
+  float percentage_Error = MAPE::apply(exact, prediction);
+  float standard_deviation = std::sqrt(RIV::apply(exact, prediction));
 
   return std::make_tuple(error, percentage_Error, standard_deviation);
 }
